Implemented distinct pair check
Included a collision check for distinct pairs only (improves performance).
Added the ability to pause the simulation while still being able to move the camera and add spawn new particles into the simulation.

Warning
Some particles fail to detect a collision and will overlap. Not sure if this was reintroduced in this particular version, but it seems more common.