Implemented automatic allocation of newly created physics objects to the physics engine and similarly for renderable objects. This removes the need for manually referring to the physics or graphics engine when creating entities from components.
Improved memory management of objects. Deleting the ProjectileManager will automatically delete projectiles stored within the vector, which in turn delete associated physics and renderable child objects. Because the child objects are deleted, they inform the physics and render engine that they are to be deleted via the observer pattern. The physics and rendering engines can then deal with the destruction of the objects by removing pointers to the object from their own vectors.