Rendering engine can now handle basic texturing of game objects. A custom directory can be defined for the location of textures (default is "./res/textures/") in the renderer class.
textures can be placed in the appropriate folder and are loaded into memory automatically when an object is spawned using the texture name as the "texture" parameter. For example, placing a texture called "asteroid.jpg" can be loaded into the engine with the texture parameter set to "asteroid.jpg".

The texture resource management is implemented so that a texture is loaded only once into the engine and is shared, using a pointer, among multiple objects. This minimises memory usage.