Implemented support for custom shaders (Still WIP). Can now specify the name of a shader file(s) through the object properties creation.
Beginning to implement a lighting system. Game objects can now own a LightSource component which automatically causes the object to cast light. Further work needed to implement support for multiple lights.

The lighting system relies on normal maps to give a perceived "depth" to the objects in the scene. As such, the lightsource itself requires a pseudo z-position to determine the direction of the light incident on a fragment. I have added the current WIP phong shader to the repository.