Added umbral shadows; shadows become progressively softer over distance of shadowed pixel from the obscuring object. There are a few bugs:
There is an inaccuracy with shadow generation for low poly objects, particular around the transition point where a new vertex is getting shaded.

The second bug is a lack of smooth interpolation in the shadow polygon trianglestrip. Need to look into how to improve this interpolation.

Final bug is some shadows having a short range, leading to undesirable lighting effects.