significantly improved shadow shading. The main issues which developed in the previous commit are mostly gone; shadow banding in particular is less pronounced. Shadow range is better. There does still exist a slight issue with shadow generation for low poly objects but these are artifacts as opposed to inaccuracies.
The main improvements comes from the inclusion of projection texture coordinates for the trapezia generated. Before, the umbral shader experienced significant distortion if the parent polygon was a high ratio trapezium. By including projection co-ordinates, this distortion is effectively mitigated.

Adding the projection co-ordinates required transferring a lot of the shadow generation algorithm to a geometry shader, which allowed for creating per-vertex w-coordinates. SFML does not support 4d texture co-ordinates, and so working around it with geometry shaders was necessary.

The remaining artifacting present is likely due to distortion along the texture's y-axis, as the trapezium essentially flattens near the transition point of visible vertices.